+++ library/src/main/java/com/getstream/sdk/chat/adapter/MessageViewHolderFactory.kt
@@ -5,6 +5,11 @@ import com.getstream.sdk.chat.adapter.MessageListItem.DateSeparatorItem
 import com.getstream.sdk.chat.adapter.MessageListItem.MessageItem
 import com.getstream.sdk.chat.adapter.MessageListItem.ThreadSeparatorItem
 import com.getstream.sdk.chat.adapter.MessageListItem.TypingItem
+import com.getstream.sdk.chat.adapter.viewholder.message.BaseMessageListItemViewHolder
+import com.getstream.sdk.chat.adapter.viewholder.message.DateSeparatorViewHolder
+import com.getstream.sdk.chat.adapter.viewholder.message.MessageListItemViewHolder
+import com.getstream.sdk.chat.adapter.viewholder.message.ThreadSeparatorViewHolder
+import com.getstream.sdk.chat.adapter.viewholder.message.TypingIndicatorViewHolder
 import com.getstream.sdk.chat.view.MessageListView
 import com.getstream.sdk.chat.view.MessageListViewStyle
 import io.getstream.chat.android.client.models.Channel
@@ -62,11 +67,9 @@ open class MessageViewHolderFactory {
                     bubbleHelper,
                     listenerContainer.messageClickListener,
                     listenerContainer.messageLongClickListener,
-                    listenerContainer.attachmentClickListener,
                     listenerContainer.reactionViewClickListener,
                     listenerContainer.userClickListener,
-                    listenerContainer.readStateClickListener,
-                    listenerContainer.giphySendListener
+                    listenerContainer.readStateClickListener
                 )
             MESSAGEITEM_TYPING ->
                 TypingIndicatorViewHolder(parent, style)