+++ library/src/main/java/com/getstream/sdk/chat/adapter/viewholder/message/BaseMessageListItemViewHolder.kt
@@ -1,8 +1,9 @@
-package com.getstream.sdk.chat.adapter
+package com.getstream.sdk.chat.adapter.viewholder.message
 
 import android.content.Context
 import android.view.View
 import androidx.recyclerview.widget.RecyclerView
+import com.getstream.sdk.chat.adapter.MessageListItem
 
 abstract class BaseMessageListItemViewHolder<T : MessageListItem>(
     itemView: View
@@ -11,18 +12,10 @@ abstract class BaseMessageListItemViewHolder<T : MessageListItem>(
     /**
      * Workaround to allow a downcast of the MessageListItem to T
      */
-    fun bindListItem(
-        messageListItem: MessageListItem,
-        position: Int
-    ) {
-        @Suppress("UNCHECKED_CAST")
-        bind(messageListItem as T, position)
-    }
+    @Suppress("UNCHECKED_CAST")
+    fun bindListItem(messageListItem: MessageListItem) = bind(messageListItem as T)
 
-    protected abstract fun bind(
-        messageListItem: T,
-        position: Int
-    )
+    protected abstract fun bind(messageListItem: T)
 
     protected val context: Context
         get() = itemView.context