+++ library/src/main/java/com/getstream/sdk/chat/adapter/viewholder/message/ThreadSeparatorViewHolder.kt
@@ -1,7 +1,8 @@
-package com.getstream.sdk.chat.adapter
+package com.getstream.sdk.chat.adapter.viewholder.message
 
 import android.view.ViewGroup
 import com.getstream.sdk.chat.adapter.MessageListItem.ThreadSeparatorItem
+import com.getstream.sdk.chat.adapter.inflater
 import com.getstream.sdk.chat.databinding.StreamItemThreadSeparatorBinding
 
 class ThreadSeparatorViewHolder(
@@ -10,10 +11,7 @@ class ThreadSeparatorViewHolder(
         StreamItemThreadSeparatorBinding.inflate(parent.inflater, parent, false)
 ) : BaseMessageListItemViewHolder<ThreadSeparatorItem>(binding.root) {
 
-    override fun bind(
-        messageListItem: ThreadSeparatorItem,
-        position: Int
-    ) {
+    override fun bind(messageListItem: ThreadSeparatorItem) {
         /* Empty */
     }
 }